static/images/menu zomato/requirements.txt
pillow
pillow-avif-plugin
tqdm
